Move ManaAbility related classes to mana package
