+++ api/src/main/java/dev/aurelium/skills/api/event/mana/ManaAbilityActivateEvent.java
@@ -1,7 +1,7 @@
 package dev.aurelium.skills.api.event.mana;
 
 import dev.aurelium.skills.api.AureliumSkillsApi;
-import dev.aurelium.skills.api.ability.ManaAbility;
+import dev.aurelium.skills.api.mana.ManaAbility;
 import dev.aurelium.skills.api.event.AureliumSkillsEvent;
 import dev.aurelium.skills.api.event.Cancellable;
 import dev.aurelium.skills.api.player.SkillsPlayer;
