+++ api/src/main/java/dev/aurelium/skills/api/mana/ManaAbility.java
@@ -1,4 +1,4 @@
-package dev.aurelium.skills.api.ability;
+package dev.aurelium.skills.api.mana;
 
 import dev.aurelium.skills.api.util.NamespacedId;
