+++ api/src/main/java/dev/aurelium/skills/api/mana/ManaAbilityProperties.java
@@ -1,5 +1,6 @@
-package dev.aurelium.skills.api.ability;
+package dev.aurelium.skills.api.mana;
 
+import dev.aurelium.skills.api.mana.ManaAbility;
 import dev.aurelium.skills.api.skill.Skill;
 import org.jetbrains.annotations.NotNull;
