+++ api/src/main/java/dev/aurelium/skills/api/skill/SkillProperties.java
@@ -2,7 +2,7 @@ package dev.aurelium.skills.api.skill;
 
 import com.google.common.collect.ImmutableList;
 import dev.aurelium.skills.api.ability.Ability;
-import dev.aurelium.skills.api.ability.ManaAbility;
+import dev.aurelium.skills.api.mana.ManaAbility;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
