+++ common/src/main/java/dev/aurelium/skills/common/api/implementation/ApiSkillsPlayer.java
@@ -2,7 +2,7 @@ package dev.aurelium.skills.common.api.implementation;
 
 
 import dev.aurelium.skills.api.ability.Ability;
-import dev.aurelium.skills.api.ability.ManaAbility;
+import dev.aurelium.skills.api.mana.ManaAbility;
 import dev.aurelium.skills.api.player.SkillsPlayer;
 import dev.aurelium.skills.api.skill.Skill;
 import dev.aurelium.skills.api.stat.Stat;
