+++ common/src/main/java/dev/aurelium/skills/common/skill/SkillDefaults.java
@@ -3,8 +3,8 @@ package dev.aurelium.skills.common.skill;
 import com.google.common.collect.ImmutableList;
 import dev.aurelium.skills.api.ability.Abilities;
 import dev.aurelium.skills.api.ability.Ability;
-import dev.aurelium.skills.api.ability.ManaAbilities;
-import dev.aurelium.skills.api.ability.ManaAbility;
+import dev.aurelium.skills.api.mana.ManaAbilities;
+import dev.aurelium.skills.api.mana.ManaAbility;
 import dev.aurelium.skills.api.skill.Skills;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
