+++ common/src/main/java/dev/aurelium/skills/common/skill/SkillRegistry.java
@@ -2,7 +2,7 @@ package dev.aurelium.skills.common.skill;
 
 import com.google.common.collect.ImmutableList;
 import dev.aurelium.skills.api.ability.Ability;
-import dev.aurelium.skills.api.ability.ManaAbility;
+import dev.aurelium.skills.api.mana.ManaAbility;
 import dev.aurelium.skills.api.skill.Skill;
 import dev.aurelium.skills.api.skill.SkillProperties;
 import dev.aurelium.skills.api.skill.Skills;
